+++ NEWS.md
@@ -5,6 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.22] 2025-06-25
+
+### Fixed
+
+* Support for `ForwardDiff.jl` v1
+
+## [0.10.21] 2025-06-24
+
+### Added
+
+* Support for `ForwardDiff.jl` v1
+
+## [0.10.20] 2025-06-16
+
+### Fixed
+
+* Fix in-place computation of `exp!` for `GeneralLinear`.
+
+## [0.10.19] 2025-06-14
+
+### Added
+
+* `riemannian_gradient` for the `GeneralLinear` manifold
+* `riemannian_gradient` method for `ProductManifold`.
+
+### Fixed
+
+* Fix the projection and embedding of tangent vectors on `GeneralLinear`.
+
 ## [0.10.18] 2025-05-29
 
 ### Fixed